Rename: Move files to view package
