+++ app/src/main/java/com/sparta/imagesearch/view/SearchFragment.kt
@@ -1,4 +1,4 @@
-package com.sparta.imagesearch
+package com.sparta.imagesearch.view
 
 import android.content.Context
 import android.os.Bundle
