+++ app/src/main/res/layout/fragment_folder.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/theme_bg"
-    tools:context=".FolderFragment">
+    tools:context=".view.FolderFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_folder"
